refactor(admin-ui): Phase2-P1-2 A4 — ScrapeTab を UrlScrapeTab.tsx へ抽出 + 未使用依存除去 (機能変更なし) (GID 1214120401966111)
